- move thumb_utils to thumbs module
